% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASD_ASTFB_mR_data}
\alias{ASD_ASTFB_mR_data}
\title{ASTFB-specific mRNA expression data in ASD samples}
\format{
ASD_ASTFB_mR_data: A matrix object with 338 samples (rows) and 15608 lncRNAs (columns).
}
\description{
ASTFB-specific mRNA expression data in ASD samples
}
\details{
The raw count data of lncRNA and mRNA expression data and meta information in ASD are obtained
from the UCSC cell browser (https://cells.ucsc.edu/?ds=autism).
}
