% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASD_Neumat_ncR_data}
\alias{ASD_Neumat_ncR_data}
\title{Neumat-specific lncRNA expression data in ASD samples}
\format{
ASD_Neumat_ncR_data: A matrix object with 524 samples (rows) and 1611 lncRNAs (columns).
}
\description{
Neumat-specific lncRNA expression data in ASD samples
}
\details{
The raw count data of lncRNA and mRNA expression data and meta information in ASD are obtained
from the UCSC cell browser (https://cells.ucsc.edu/?ds=autism).
}
