% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASTFB_ASD_lncRNAs_data}
\alias{ASTFB_ASD_lncRNAs_data}
\title{ASTFB-specific lncRNA expression data in ASD samples}
\format{
ASTFB_ASD_lncRNAs_data: A matrix object with 2448 samples (columns) and 813 lncRNAs (rows).
}
\description{
ASTFB-specific lncRNA expression data in ASD samples
}
\details{
The normalized data of lncRNA and mRNA expression data and meta information in ASD are obtained
from the UCSC cell browser (https://cells.ucsc.edu/?ds=autism).
}
