% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{csn_edge}
\alias{csn_edge}
\title{Cell-specific network of edge gx-gy}
\usage{
csn_edge(
  gx,
  gy,
  boxsize = 0.1,
  dev = TRUE,
  iteration = TRUE,
  cell_id = NULL,
  maxiter = 20
)
}
\arguments{
\item{gx}{Gene expression values of gene x (a vector) in n cells}

\item{gy}{Gene expression values of gene y (a vector) in n cells}

\item{boxsize}{Size of neighborhood (0.1 in default)}

\item{dev}{Whether to use local standard deviation (TRUE in default)}

\item{iteration}{Whether need iterate (TRUE in default)}

\item{cell_id}{cell numbers, NULL means all cell to compute (NULL in default)}

\item{maxiter}{Maximum Number Of Iterations (20 in default)}
}
\value{
res a vector, the normalized statistic of edge gx-gy in n cells
}
\description{
Cell-specific network of edge gx-gy
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
Dai H, Li L, Zeng T, Chen L. Cell-specific network constructed by single-cell RNA sequencing data. Nucleic Acids Res. 2019 Jun 20;47(11):e62. doi: 10.1093/nar/gkz172.
}
