% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Cycle_network}
\alias{Cycle_network}
\title{Constructing cell type-specific lncRNA-mRNA regulatory network}
\usage{
Cycle_network(
  lncR,
  mR,
  boxsize = 0.1,
  p.value.cutoff = 0.05,
  num.cores = 2,
  dev = TRUE,
  iteration = TRUE,
  cell_id = NULL,
  maxiter = 20
)
}
\arguments{
\item{lncR}{Gene expression values of lncRNAs in single cells, rows are cells and columns are lncRNAs}

\item{mR}{Gene expression values of mRNAs in single cells, rows are cells and columns are mRNAs}

\item{boxsize}{Size of neighborhood (0.1 in default)}

\item{p.value.cutoff}{cutoff of p.value (0.05 in default)}

\item{num.cores}{number of parallel cores (2 in default)}

\item{dev}{Whether to use local standard deviation (TRUE in default)}

\item{iteration}{Whether need iterate (TRUE in default)}

\item{cell_id}{cell numbers, NULL means all cell to compute (NULL in default)}

\item{maxiter}{Maximum Number Of Iterations (20 in default)}
}
\value{
res_list: a list of cell type-specific lncRNA-mRNA regulatory network
}
\description{
Constructing cell type-specific lncRNA-mRNA regulatory network
}
\examples{
# NOT RUN
# load('ASD_exp_3cell_types.rda')
# microglia_Cycle_networks <- Cycle_network(Microglia_ASD_lncRNAs_data[1:100,1:100], Microglia_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# ASTFB_Cycle_networks <- Cycle_network(ASTFB_ASD_lncRNAs_data[1:100,1:100],ASTFB_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Neumat_Cycle_networks <- Cycle_network(Neumat_ASD_lncRNAs_data[1:100,1:100],Neumat_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
