% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microglia_ASD_mRNAs_data}
\alias{Microglia_ASD_mRNAs_data}
\title{microgila-specific mRNA expression data in ASD samples}
\format{
Microglia_ASD_mRNAs_data: A matrix object with 1509 samples (columns) and 5133 mRNAs (rows).
}
\description{
microgila-specific mRNA expression data in ASD samples
}
\details{
The normalized data of lncRNA and mRNA expression data and meta information in ASD are obtained
from the UCSC cell browser (https://cells.ucsc.edu/?ds=autism).
}
