% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Overlap.net}
\alias{Overlap.net}
\title{Overlap.ceRNet}
\usage{
Overlap.net(net, overlap.num = 1, type = c("equal", "least"))
}
\arguments{
\item{net}{List object, the list of networks.}

\item{overlap.num}{The minimum number of interactions existing in multiple networks.}

\item{type}{The overlapped interactions in overlap.num networks ("equal") or at least overlap.num networks ("least").}
}
\value{
Matrix object: The overlapped interactions.
}
\description{
Identifying the overlap between multiple networks.
}
\examples{
library(igraph)
net <- list(as_data_frame(sample_k_regular(10, 2)), as_data_frame(sample_k_regular(10, 3)), as_data_frame(sample_k_regular(10, 2)), as_data_frame(sample_k_regular(10, 3)))
ceRNet_2 <- Overlap.ceRNet(net, overlap.num = 2, type = "least")

}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
