% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Random_net_parallel}
\alias{Random_net_parallel}
\title{Generate random networks in parallel and statistical significance p value of topological characteristics (path length and density) in biological networks.}
\usage{
Random_net_parallel(
  obser_path,
  obser_density,
  nodes.num,
  edges.num,
  perm = 100,
  directed = FALSE,
  num.cores = 2
)
}
\arguments{
\item{obser_path}{Observed characteristics path length of a biological network.}

\item{obser_density}{Observed density of a biological network.}

\item{nodes.num}{the number of nodes}

\item{edges.num}{the number of edges}

\item{perm}{The number of permutations for generating random networks.}

\item{directed}{a logical object, TRUE or FALSE means directed or undirected network, respectively.}

\item{num.cores}{number of parallel cores (2 in default)}
}
\value{
A vector: The significance -log10(p-value) of topological characteristics (path length and density) of a biological network.
}
\description{
Generate random networks in parallel and statistical significance p value of topological characteristics (path length and density) in biological networks.
}
