% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{upperlower}
\alias{upperlower}
\title{Define the upper and lower neighborhood of each plot}
\usage{
upperlower(gx, gy, boxsize = 0.1)
}
\arguments{
\item{gx}{Gene expression values of gene x (a vector) in n cells}

\item{gy}{Gene expression values of gene y (a vector) in n cells}

\item{boxsize}{Size of neighborhood (0.1 in default)}
}
\value{
upperlower_result: a list of upper result and lower result
}
\description{
Define the upper and lower neighborhood of each plot
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
Dai H, Li L, Zeng T, Chen L. Cell-specific network constructed by single-cell RNA sequencing data. Nucleic Acids Res. 2019 Jun 20;47(11):e62. doi: 10.1093/nar/gkz172.
}
