% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{upperlower_dev}
\alias{upperlower_dev}
\title{Define the upper and lower neighborhood of each plot using local standard deviation}
\usage{
upperlower_dev(
  gene1,
  gene2,
  boxsize = 0.1,
  iteration = TRUE,
  cell_id = NULL,
  maxiter = 20
)
}
\arguments{
\item{gene1}{Gene expression values of gene1 (a vector) in n cells}

\item{gene2}{Gene expression values of gene2 (a vector) in n cells}

\item{boxsize}{Size of neighborhood (0.1 in default)}

\item{iteration}{Whether need iterate (TRUE in default)}

\item{cell_id}{cell numbers, NULL means all cell to compute (NULL in default)}

\item{maxiter}{Maximum Number Of Iterations (20 in default)}
}
\value{
upperlower_devresult: a list of upper result and lower result
}
\description{
Define the upper and lower neighborhood of each plot using local standard deviation
}
