% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Overlap.hub}
\alias{Overlap.hub}
\title{Overlap.hub}
\usage{
Overlap.hub(hub, overlap.num = 1, type = c("equal", "least"))
}
\arguments{
\item{hub}{List object, the list of hubs.}

\item{overlap.num}{The minimum number of hubs existing in multiple lists of hubs}

\item{type}{The overlapped hubs in overlap.num hub lists ("equal") or at least overlap.num hub lists ("least").}
}
\value{
A vector: The overlapped hubs.
}
\description{
Identifying the overlap between multiple lists of hubs.
}
\examples{
# NOT RUN
# load('ASD_exp_3cell_types.rda')
# microglia_Cycle_networks <- Cycle_network(Microglia_ASD_lncRNAs_data[1:100,1:100], Microglia_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# ASTFB_Cycle_networks <- Cycle_network(ASTFB_ASD_lncRNAs_data[1:100,1:100],ASTFB_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Neumat_Cycle_networks <- Cycle_network(Neumat_ASD_lncRNAs_data[1:100,1:100],Neumat_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Cycle_networks = list(microglia_Cycle_networks, ASTFB_Cycle_networks, Neumat_Cycle_networks)
# Cycle_hubs_lncRNAs <- hub_discovery(Cycle_networks)
# Overlap_hub.2 <- Overlap.hub(Cycle_hubs_lncRNAs, overlap.num = 2, type = "least")

}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
