% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Overlap.net}
\alias{Overlap.net}
\title{Overlap.net}
\usage{
Overlap.net(net, overlap.num = 1, type = c("equal", "least"))
}
\arguments{
\item{net}{List object, the list of networks.}

\item{overlap.num}{The minimum number of interactions existing in multiple networks.}

\item{type}{The overlapped interactions in overlap.num networks ("equal") or at least overlap.num networks ("least").}
}
\value{
Matrix object: The overlapped interactions.
}
\description{
Identifying the overlap between multiple networks.
}
\examples{
# NOT RUN
# load('ASD_exp_3cell_types.rda')
# microglia_Cycle_networks <- Cycle_network(Microglia_ASD_lncRNAs_data[1:100,1:100], Microglia_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# ASTFB_Cycle_networks <- Cycle_network(ASTFB_ASD_lncRNAs_data[1:100,1:100],ASTFB_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Neumat_Cycle_networks <- Cycle_network(Neumat_ASD_lncRNAs_data[1:100,1:100],Neumat_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Cycle_networks = list(microglia_Cycle_networks, ASTFB_Cycle_networks, Neumat_Cycle_networks)
# Overlap_net.all <- Overlap.net(Cycle_networks, overlap.num = 1, type = "least")
# Overlap_net.1 <- Overlap.net(Cycle_networks, overlap.num = 1, type = "equal")

}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
