% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Random_net_parallel}
\alias{Random_net_parallel}
\title{Generate random networks in parallel and statistical significance p value of topological characteristics (path length and density) in biological networks.}
\usage{
Random_net_parallel(
  obser_path,
  obser_density,
  nodes.num,
  edges.num,
  perm = 100,
  directed = FALSE,
  num.cores = 2
)
}
\arguments{
\item{obser_path}{Observed characteristics path length of a biological network.}

\item{obser_density}{Observed density of a biological network.}

\item{nodes.num}{the number of nodes}

\item{edges.num}{the number of edges}

\item{perm}{The number of permutations for generating random networks.}

\item{directed}{a logical object, TRUE or FALSE means directed or undirected network, respectively.}

\item{num.cores}{number of parallel cores (2 in default)}
}
\value{
A List: The significance p-value and -log10(p-value) of topological characteristics (path length and density) of a biological network.
}
\description{
Generate random networks in parallel and statistical significance p value of topological characteristics (path length and density) in biological networks.
}
\examples{
#' # NOT RUN
# load('ASD_exp_3cell_types.rda')
# Cycle_networks <- Cycle_network(ASD_Neumat_ncR_data[1:100,1:100], ASD_Neumat_mR_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# g1 <- Cycle_networks
# g1 <- make_graph(c(t(g1[, 1:2])),directed = F)
# m <- gsize(g1)
# n <- vcount(g1)
# l <- mean_distance(g1)
# d <- edge_density(g1)
# pvalue_path_density <- Random_net_parallel(l, d, n, m, perm = 100, directed = FALSE, num.cores = 4)

}
\references{
Csardi G, Nepusz T. The igraph software package for complex network research.
InterJournal, Complex Systems. 2006, 1695.
}
