% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Sim.hub}
\alias{Sim.hub}
\title{Calculating similarity matrix between two list of hubs}
\usage{
Sim.hub(hub1, hub2)
}
\arguments{
\item{hub1}{List object, the first list of hub}

\item{hub2}{List object, the second list of hub}
}
\value{
Sim: a similarity matrix between two list of hubs
}
\description{
Calculating similarity matrix between two list of hubs
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
