% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Sim.network}
\alias{Sim.network}
\title{Calculating similarity matrix between two list of cell type-specific lncRNA-mRNA networks}
\usage{
Sim.network(net1, net2, directed = TRUE)
}
\arguments{
\item{net1}{List object, the first list of network}

\item{net2}{List object, the second list of network}

\item{directed}{Logical value, network directed (TRUE) or undirected (FALSE)}
}
\value{
Sim: a similarity matrix between two list of networks
}
\description{
Calculating similarity matrix between two list of cell type-specific lncRNA-mRNA networks
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
