% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{Sim.network}
\alias{Sim.network}
\title{Calculating similarity matrix between two list of cell type-specific lncRNA-mRNA networks}
\usage{
Sim.network(net1, net2, directed = TRUE)
}
\arguments{
\item{net1}{List object, the first list of network}

\item{net2}{List object, the second list of network}

\item{directed}{Logical value, network directed (TRUE) or undirected (FALSE)}
}
\value{
A Matrix: a similarity matrix between two list of networks
}
\description{
Calculating similarity matrix between two list of cell type-specific lncRNA-mRNA networks
}
\examples{
# NOT RUN
# load('ASD_exp_3cell_types.rda')
# microglia_Cycle_networks <- Cycle_network(Microglia_ASD_lncRNAs_data[1:100,1:100], Microglia_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# ASTFB_Cycle_networks <- Cycle_network(ASTFB_ASD_lncRNAs_data[1:100,1:100],ASTFB_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Neumat_Cycle_networks <- Cycle_network(Neumat_ASD_lncRNAs_data[1:100,1:100],Neumat_ASD_mRNAs_data[1:100,1:100], boxsize = 0.1, p.value.cutoff = 0.05, num.cores = 2, dev = TRUE, iteration = TRUE, cell_id = NULL, maxiter = 20)
# Cycle_networks = list(microglia_Cycle_networks, ASTFB_Cycle_networks, Neumat_Cycle_networks)
# Cycle_network_Sim <- Sim.network(Cycle_networks, Cycle_networks, directed = TRUE)

}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.,22(1):578.
}
