% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle.R
\name{upperlower}
\alias{upperlower}
\title{Define the upper and lower neighborhood of each plot}
\usage{
upperlower(gx, gy, boxsize = 0.1)
}
\arguments{
\item{gx}{Gene expression values of gene x (a vector) in n cells}

\item{gy}{Gene expression values of gene y (a vector) in n cells}

\item{boxsize}{Size of neighborhood (0.1 in default)}
}
\value{
upperlower_result: a list of upper result and lower result
}
\description{
Define the upper and lower neighborhood of each plot
}
